+++ src/entities/Hero.cpp
@@ -109,6 +109,7 @@ Hero::Hero(Equipment& equipment):
   end_invincible_date(0),
   normal_walking_speed(88),
   walking_speed(normal_walking_speed),
+  carry_height(18),
   delayed_teletransporter(nullptr),
   on_raised_blocks(false),
   last_solid_ground_coords(0, 0),
@@ -116,7 +117,8 @@ Hero::Hero(Equipment& equipment):
   target_solid_ground_callback(),
   next_ground_date(0),
   next_ice_date(0),
-  ice_movement_direction8(0) {
+  ice_movement_direction8(0),
+  push_delay(800) {
 
   // position
   set_origin(8, 13);
@@ -984,6 +986,43 @@ void Hero::set_walking_speed(int walking_speed) {
     this->walking_speed = walking_speed;
     get_state()->notify_walking_speed_changed();
   }
+}
+ 	
+/**
+* \brief Returns the default height carried objects will be displayed at.
+* \return The height in pixels.
+*/
+int Hero::get_carry_height() const{
+  return carry_height;
+}
+
+/**
+* \brief Sets the default height carried objects will be displayed at.
+* \param carry_height The height in pixels.
+  */
+void Hero::set_carry_height(int carry_height) {
+  std::shared_ptr<CarriedObject> carried = get_carried_object();
+
+  this->carry_height = carry_height;
+
+  if (carried != nullptr) {
+    carried->update_relative_movement();
+  }
+}
+	
+/**
+* \brief Returns the delay between the moment the hero start pushing on an obstacle, and the moment he actually enters the push state.
+*/
+int Hero::get_push_delay() const {
+  return push_delay;
+}
+
+/**
+ * \brief Sets the delay between the moment the hero start pushing on an obstacle, and the moment he actually enters the push state.
+ * \param push_delay The new delay in ms.
+ */
+void Hero::set_push_delay(int push_delay) {
+  this->push_delay = push_delay;
 }
 
 /**
@@ -2127,11 +2166,11 @@ void Hero::notify_grabbed_entity_collision() {
  * hero wants to cut a bush or some grass.
  * Returns false by default.
  *
- * \param entity The entity to check.
+ * \param destructible The entity to check.
  * \return \c true if the sword is cutting this entity.
  */
-bool Hero::is_striking_with_sword(Entity& entity) const {
-  return get_state()->is_cutting_with_sword(entity);
+bool Hero::is_cutting_with_sword(Destructible& destructible) const {
+  return get_state()->is_cutting_with_sword(destructible);
 }
 
 /**